extract date & venue from content blocks in frontend view #77
