+++ src/routes/create/DateAndVenue/index.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import AddDate from '../addDate'
+
+const DateAndVenue = () => {
+  return (
+    <AddDate />
+  )
+}
+export default DateAndVenue
